src/bot/requirements.txt
openai~=1.23.2
peewee~=3.16.3
six~=1.16.0
python-dotenv~=1.0.0
gspread~=5.12.0
dspy-ai[groq]~=2.4.9
google-cloud-firestore~=2.16.0
functions-framework~=3.7.0
coinbase-advanced-py~=1.4.2
